HARMONY-1172: Clean up code locations, naming
