+++ app/models/services/turbo-service.ts
@@ -1,6 +1,5 @@
 import _ from 'lodash';
 import { Logger } from 'winston';
-import { Job } from '../job';
 import BaseService from './base-service';
 import InvocationResult from './invocation-result';
 
@@ -22,21 +21,5 @@ export default class TurboService extends BaseService<TurboServiceParams> {
    */
   async _run(_logger: Logger): Promise<InvocationResult> {
     return null;
-  }
-}
-
-/**
- * Extends TurboService for testing purposes.
- */
-export class TestTurboService extends TurboService {
-  /**
-   * Calls _createJob which creates a job from a service and its operation.
-   * @param requestUrl - The URL the end user invoked
-   * @returns The created job
-   */
-  createJob(
-    requestUrl: string,
-  ): Job {
-    return this._createJob(requestUrl);
   }
 }
