+++ test/helpers/jobs.ts
@@ -5,7 +5,7 @@ import { v4 as uuid } from 'uuid';
 import { Application } from 'express';
 import _ from 'lodash';
 import JobLink from '../../app/models/job-link';
-import { Job, JobStatus, JobRecord, jobRecordFields, SerializedJob, getRelatedLinks } from '../../app/models/job';
+import { Job, JobStatus, JobRecord, jobRecordFields, JobForDisplay, getRelatedLinks } from '../../app/models/job';
 import { JobListing } from '../../app/frontends/jobs';
 import db, { Transaction } from '../../app/util/db';
 import { hookRequest } from './hooks';
@@ -143,7 +143,7 @@ export function areStacJobLinksEqual(jobLinks: JobLink[], stacLinks: JobLink[]):
  */
 export function jobsEqual(
   jobRecord: JobRecord,
-  serializedJob: SerializedJob,
+  serializedJob: JobForDisplay,
   skipLinks = false,
   skipMessage = false): boolean {
   const recordLinks = new Job(jobRecord).getRelatedLinks('data');
